requirements.txt
discord.py
requests
jaconv
unidecode
rapidfuzz
python-dotenv
flask